backend/requirements.txt
protobuf==4.25.3
grpcio==1.62.1
grpcio-tools==1.62.1
SQLAlchemy==2.0.25
python-dotenv==1.0.1